Rewrite of the AuthenticationApplication (v0.2.0.0 now)

+++ SunDofus.Realm/Database/Cache/MapsCache.cs
@@ -23,7 +23,7 @@ namespace realm.Database.Cache
                 {
                     var myMap = new Models.Maps.MapModel();
 
-                    myMap.myId = SQLReader.GetInt16("id");
+                    myMap.myId = SQLReader.GetInt32("id");
                     myMap.myDate = SQLReader.GetInt32("date");
                     myMap.myWidth = SQLReader.GetInt16("width");
                     myMap.myHeight = SQLReader.GetInt16("heigth");
